+++ CHANGELOG.md
@@ -12,6 +12,8 @@ speaking, considered to be the player's experience.
 
 ### Added
 
+- Automatically copy the map ID to the clipboard when opening the in-game menu,
+  allowing for easy copy-pasting of the map ID.
 - Added a What's New screen, shown on the first startup of each new version.
 
 ### Fixed